requirements.txt
nbformat>=5.7.0
jupyter>=1.1.0
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
openai>=1.0.0
python-dotenv>=0.17.0
